refactor: Updated file structure
