+++ src/routes/graphql/schema/profile/profilesQuery.ts
@@ -5,7 +5,7 @@ import {
   ThunkObjMap,
 } from "graphql";
 
-import { GraphQlContext } from "../context";
+import { GraphQlContext } from "../../context";
 import { ProfileType } from "./profilesType";
 
 type ProfileArgs = {
